backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
ytmusicapi==1.7.5
pydantic==2.8.2
python-dotenv==1.0.1
pytest==8.2.2
pytest-asyncio==0.23.7
anyio==4.4.0
google-auth==2.34.0
google-genai==1.0.0
